quickfix

+++ tafelServer/ServerRequestHandler.java
@@ -76,7 +76,7 @@ public class ServerRequestHandler {
 		String antwort = "Nachricht mit ID=" + deleteRequest.getMessageID() + " gelöscht!";
 		try {
 			Message message = anzeigetafel.getMessageByID(deleteRequest.getMessageID());
-			if (message.isOeffentlich()
+			if ((message.isOeffentlich() &&  message.getAbtNr()==anzeigetafel.getAbteilungsID())
 					&& (message.getUserID() == deleteRequest.getUserID() || deleteRequest.getUserID() == 1))
 				tafelServer.deletePublicMessage(deleteRequest.getMessageID());
 
@@ -126,7 +126,7 @@ public class ServerRequestHandler {
 		String antwort = "Nachricht mit ID=" + modifyRequest.getMessageID() + " geändert!";
 		try {
 			Message message = anzeigetafel.getMessageByID(modifyRequest.getMessageID());
-			if (message.isOeffentlich()
+			if ((message.isOeffentlich() && message.getAbtNr()==anzeigetafel.getAbteilungsID())
 					&& (message.getUserID() == modifyRequest.getUserID() || modifyRequest.getUserID() == 1))
 				tafelServer.modifyPublicMessage(modifyRequest.getMessageID(), modifyRequest.getNewMessage());
